At --start-at NUMBER, --end-at NUMBER
e.g.

```
$ ./bin/printable-ascii --start-at 65 --end-at 71
A
B
C
D
E
F
G
```

The start/end will be limited to 32-126. If you pick a number greater
than 126 it will be changed to 126. If you pick a number less than 32 it
will be changed to 32.

If you pick an end less than the start then you get no output.