Allow --start-at / --end-at to accept ASCII
If given a single ASCII character then `--start-at` / `--end-at` will
start/end the list at the base10 value of that character.

```
$ printable-ascii --start-at A --end-at C
A
B
C
```

```
$ printable-ascii --start-at 65 --end-at 67
A
B
C
```

To start at a punctuation value reliably you must place it in double
quotes due to shell behavior.

```
$ bin/printable-ascii --start-at " " --end-at "(" --decimal --json | jq -c '.[]'
{"character":" ","decimal":"32"}
{"character":"!","decimal":"33"}
{"character":"\"","decimal":"34"}
{"character":"#","decimal":"35"}
{"character":"$","decimal":"36"}
{"character":"%","decimal":"37"}
{"character":"&","decimal":"38"}
{"character":"'","decimal":"39"}
{"character":"(","decimal":"40"}
```

This is a breaking API change!

Prior to this a single digit value would be converted to 32 as all
start/end numbers outside of the printable range are automatically
converted to their closest value in range.

Now a single digit value won't be converted to 32 (the closest printable
value) but be treated as a single ASCII character.

```
$ printable-ascii --start-at 0 --end-at 9 --decimal
CHARACTER   DECIMAL
0           48
1           49
2           50
3           51
4           52
5           53
6           54
7           55
8           56
9           57
```