Refine parsing of start/end and ranges
BREAKING CHANGE: values outside of 32-126 will no longer be silently
converted to 32 (if lower than) or 126 (if greater than). They will now
be an error exit.

Add clarification to STDERR when a range will output no characters which
can be surprising, e.g. 1-33 starts with ASCII "1" and so parses as
the range 49-33.