Add the `--times NUMBER` option
This option repeats the given output `NUMBER` times.

There's more code changing here than you may expect. That's largely
around the idea that with `--times` and `--random` we want the
randomness to continue to be, well, random as it repeats the pattern.

e.g. we want

```
$ ./bin/printable-ascii --hex-color --times 4
EB0726
B7D0F4
CAA45B
CAC258
```

and not

```
$ ./bin/printable-ascii --hex-color --times 4
EB0726
EB0726
EB0726
EB0726
```

With these changes, we have it! All the outputs now accomodate both
random and times options.