Now working in Docker with STDIN filtering
Without this code adjustment the script was not runnable in docker
without forcing the end user to supply `-t` to run the container with a
TTY. I wanted to keep the simplicity so took on some code complexity.

```
docker run sdball/printable-ascii
```

After a bunch of experimentation with what's possible using $stdin and
IO I landed on ultimately trying to read a byte from STDIN with a
nonblocking call. If there is nothing to read from STDIN then an
exception is raised that basically says "this call would block waiting
for STDIN input" which I catch and have the script behave as normal.

To filter STDIN from the Docker image unavoidably requires using the
`-i` flag for the docker run.

```
$ echo "xyzzy ⌥pl⌃ugh" | docker run -i printable-ascii:local
xyzzy plugh
```

Passing options to the script works as it should

```
$ echo "xyzzy ⌥pl⌃ugh" | docker run -i printable-ascii:local --range x-z
xyzzy
```