Include ASCII control characters
Yes control characters are not printable ASCII. BUT we can allow them to
be specially included in both the ASCII informational output and in the
STDIN filtering.

When included in ASCII info output the control characters will be
represented by their abbreviations e.g. NUL instead of the literal null
character.

When included in STDIN filtering the control characters will be
translated to printable ASCII e.g. <NUL>. The only exception is the EOL
(newline) character which will be literally preserved if the `--newline`
option is given.